+++ scripts/lint-allow-test-rule-refs.ceiling.json
@@ -1,4 +1,4 @@
 {
-  "maxFiles": 314,
+  "maxFiles": 309,
   "grace": 3
 }
