+++ scripts/lint-test-file-count.allowlist.json
@@ -146,14 +146,6 @@
         "prompt-budget.test.cjs"
       ],
       "issue": "2929"
-    },
-    "audit": {
-      "files": [
-        "audit-command-cutover.test.cjs",
-        "audit-fix-command.test.cjs",
-        "fix-2766-audit-uat-archived-and-table-shapes.test.cjs"
-      ],
-      "issue": "2766"
     }
   }
 }
